requirements.txt
# Polymarket SDK
py-clob-client>=0.17.0
# polymarket-agents is optional; uncomment if pinned version is needed
# polymarket-agents

# LLM
anthropic>=0.40.0

# News & external data
tavily-python>=0.5.0
httpx>=0.27.0
requests>=2.32.0

# Math / data
numpy>=1.26.0
pandas>=2.2.0

# Scheduler
apscheduler>=3.10.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.6.0
loguru>=0.7.2

# Test
pytest>=8.0.0
pytest-asyncio>=0.23.0
